Move module renderers into renderers package
