--- src/heart/modules/README.md
@@ -1,1 +1,0 @@
-Self-contained visual modules and demos. Each subpackage bundles assets and logic for a specific experience (for example, Mario scenes or cellular automata) that can be scheduled by the runtime.
