--- src/heart/modules/__init__.py
@@ -1,1 +1,0 @@
-"""Shared renderers bundled as modules with providers and assets."""
