+++ src/heart/programs/configurations/lib_2025.py
@@ -5,10 +5,6 @@ from lagom import Singleton
 
 from heart.display.color import Color
 from heart.environment import GameLoop
-from heart.modules.life.renderer import Life
-from heart.modules.mario.provider import MarioRendererProvider
-from heart.modules.mario.renderer import MarioRenderer
-from heart.modules.water_cube.renderer import WaterCube
 from heart.navigation import ComposedRenderer, MultiScene
 from heart.peripheral.providers.acceleration import AllAccelerometersProvider
 from heart.renderers.artist import ArtistScene
@@ -17,8 +13,11 @@ from heart.renderers.heart_title_screen import HeartTitleScreen
 from heart.renderers.hilbert_curve import HilbertScene
 from heart.renderers.image import RenderImage
 from heart.renderers.kirby import KirbyScene
+from heart.renderers.life.renderer import Life
 from heart.renderers.mandelbrot.scene import MandelbrotMode
 from heart.renderers.mandelbrot.title import MandelbrotTitle
+from heart.renderers.mario.provider import MarioRendererProvider
+from heart.renderers.mario.renderer import MarioRenderer
 from heart.renderers.multicolor import (MulticolorRenderer,
                                         MulticolorStateProvider)
 from heart.renderers.random_pixel import RandomPixel
@@ -27,6 +26,7 @@ from heart.renderers.spritesheet_random import SpritesheetLoopRandom
 from heart.renderers.text import TextRendering
 from heart.renderers.three_fractal import FractalScene
 from heart.renderers.tixyland import Tixyland, TixylandStateProvider
+from heart.renderers.water_cube.renderer import WaterCube
 from heart.renderers.water_title_screen import WaterTitleScreen
