+++ src/heart/renderers/README.md
@@ -1,0 +1,1 @@
+Self-contained visual renderers and demos. Each subpackage bundles assets and logic for a specific experience (for example, Mario scenes or cellular automata) that can be scheduled by the runtime.
