+++ src/heart/renderers/life/__init__.py
@@ -1,4 +1,4 @@
-from heart.modules.life.provider import LifeStateProvider
 from heart.peripheral.core.providers import container
+from heart.renderers.life.provider import LifeStateProvider
 
 container[LifeStateProvider] = LifeStateProvider
