+++ src/heart/renderers/life/provider.py
@@ -4,10 +4,10 @@ from typing import Callable
 import numpy as np
 import reactivex
 
-from heart.modules.life.state import LifeState
 from heart.peripheral.core.manager import PeripheralManager
 from heart.peripheral.providers.switch import MainSwitchProvider
 from heart.peripheral.uwb import ops
+from heart.renderers.life.state import LifeState
 
 
 class LifeStateProvider:
