+++ src/heart/renderers/water_cube/provider.py
@@ -2,10 +2,10 @@
 import reactivex
 from reactivex import operators as ops
 
-from heart.modules.water_cube.state import WaterCubeState
 from heart.peripheral.core.providers import ObservableProvider
 from heart.peripheral.providers.acceleration import AllAccelerometersProvider
 from heart.peripheral.sensor import Acceleration
+from heart.renderers.water_cube.state import WaterCubeState
 
 
 class WaterCubeStateProvider(ObservableProvider[WaterCubeState]):
